ON-959 code and test complete
