+++ src/SFA.DAS.AssessorService.Application/Interfaces/IRegisterQueryRepository.cs
@@ -11,26 +11,16 @@ namespace SFA.DAS.AssessorService.Application.Interfaces
         Task<IEnumerable<DeliveryArea>> GetDeliveryAreas();
         Task<EpaOrganisation> GetEpaOrganisationById(Guid id);
         Task<EpaOrganisation> GetEpaOrganisationByOrganisationId(string organisationId);
-        Task<bool> EpaOrganisationExistsWithOrganisationId(string organisationId);
-        Task<bool> EpaOrganisationExistsWithUkprn(long ukprn);
-        Task<bool> OrganisationTypeExists(int organisationTypeId);
-        Task<bool> EpaOrganisationAlreadyUsingUkprn(long ukprn, string organisationId);
-        Task<bool> EpaOrganisationAlreadyUsingName(string organisationName, string organisationId);
+
+
         Task<string> EpaOrganisationIdCurrentMaximum();
-
         Task<int> EpaContactUsernameHighestCounter();
-
-        Task<bool> ContactIdIsValid(string contactId);
-
-        Task<bool> EmailAlreadyPresentInAnotherOrganisation(string email, string organisationId);
-        Task<bool> ContactIdIsValidForOrganisationId(string contactId, string organisationId);
         Task<IEnumerable<AssessmentOrganisationSummary>> GetAssessmentOrganisations();
-        Task<bool> EpaOrganisationStandardExists(string organisationId, int standardCode);
         Task<IEnumerable<EpaOrganisation>> GetAssessmentOrganisationsByStandardId(int standardId);
         Task<IEnumerable<AssessmentOrganisationContact>> GetAssessmentOrganisationContacts(string organisationId);
         Task<AssessmentOrganisationContact> GetPrimaryOrFirstContact(string organisationId);
         Task<IEnumerable<OrganisationStandardSummary>> GetOrganisationStandardByOrganisationId(string organisationId);
-        Task<IEnumerable<OrganisationStandardPeriod>> GetOrganisatonStandardPeriodsByOrganisationStandard(string organisationId, int standardId);
+        Task<IEnumerable<OrganisationStandardPeriod>> GetOrganisationStandardPeriodsByOrganisationStandard(string organisationId, int standardId);
         Task<IEnumerable<AssessmentOrganisationSummary>> GetAssessmentOrganisationsByUkprn(string ukprn);
         Task<IEnumerable<AssessmentOrganisationSummary>> GetAssessmentOrganisationsByOrganisationId(string organisationId);
         Task<IEnumerable<AssessmentOrganisationSummary>> GetAssessmentOrganisationsbyName(string organisationName);
